deployer/requirements.txt
celery[redis]==5.3.6
docker==6.1.3
python-dotenv==1.0.0
requests==2.31.0
pyclamd==0.4.1
